Sort the java files into respective packages
