+++ src/main/java/commands/UnmarkCommand.java
@@ -1,8 +1,16 @@
+package commands;
+
+import exception.PrimoException;
+import storage.Storage;
+import tasks.Task;
+import tasks.TaskList;
+import ui.Ui;
+
 import java.util.ArrayList;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
     int index;
-    public DeleteCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
     public boolean isExit() {
@@ -14,10 +22,10 @@ public class DeleteCommand extends Command {
         if (index >= list.size() || index + 1 <= 0) {
             throw new PrimoException("Please select within the indexes of the tasklist!");
         }
+        list.get(index).markAsUndone();
         System.out.println("\nEl Primo:");
-        System.out.println("Noted. I've removed this task:");
+        System.out.println("OK, I've marked this task as not done yet:");
         System.out.println(list.get(index));
-        list.remove(index);
         storage.updateStorage(tasks);
     }
 }
