+++ src/main/java/tasks/EventTask.java
@@ -1,3 +1,5 @@
+package tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
