+++ src/main/java/tasks/ToDoTask.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public class ToDoTask extends Task {
     private final String SYMBOL = "T";
